requirements.txt
thrift==0.8.0
mock
facebook_scribe
requests
